refactor(docs): make scroll-to-top button a component
